pyproject.toml
[project]
name = "skymail"
version = "0.1.0"
description = ""
authors = [
    {name = "preetsinghmakkar",email = "preetsinghmakkar154@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0)",
    "alembic (>=1.18.1,<2.0.0)",
    "pydantic (>=2.0)",
    "pydantic[email] (>=2.0)",
    "email-validator (>=2.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "boto3 (>=1.42.34,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
